+++ portfolio/src/main/webapp/index.html
@@ -61,8 +61,6 @@
         <input type="submit" value="Submit Your Choice" onchange="loadAndShowData()">
       </form>
 
-
-
       <div id="comments"></div>
     </div>
   </body>
